Move bulk of components to scenes
