+++ index.html
@@ -24,6 +24,6 @@
     <menu-scene></menu-scene>
     <game-scene></game-scene>
 
-    <script type="module" src="/src/bootstrap.ts"></script>
+    <script type="module" src="/src/app/app.ts"></script>
   </body>
 </html>
