+++ package.json
@@ -30,6 +30,7 @@
     "eslint-plugin-import": "^2.24.2",
     "jest": "^27.2.0",
     "prettier": "^2.4.1",
+    "sass": "^1.42.1",
     "ts-jest": "^27.0.5",
     "typescript": "^4.4.3",
     "vite": "^2.5.8"
