+++ src/ui/styles.menu.scss
@@ -1,0 +1,80 @@
+.options {
+  display: flex;
+  flex-direction: row;
+  max-width: 80%;
+
+  > {
+    display: flex;
+    flex-direction: column;
+  }
+}
+
+.toggle {
+    width:100%;
+    margin-bottom:10px;
+  position: relative;
+  display: inline-block;
+
+  * {
+    box-sizing: border-box;
+    &:before,
+    &:after {
+      content: "";
+      position: absolute;
+    }
+  }
+
+  input {
+    height: 40px;
+    left: 0;
+    opacity: 0;
+    position: absolute;
+    top: 0;
+    width: 40px;
+  }
+
+  label.toggle-item {
+    width: 3em;
+    height: 1.5em;
+    display: inline-block;
+    border-radius: 50px;
+    position: relative;
+    transition: all 0.2s ease;
+    transform-origin: 20% center;
+    cursor: pointer;
+    background: #af4c4c;
+    border: 0.5px solid rgba(117, 117, 117, 0.31);
+    box-shadow: inset 0px 0px 4px 0px rgba(0, 0, 0, 0.2),
+      0 -3px 4px rgba(0, 0, 0, 0.15);
+    &:before {
+      display: block;
+      top: 0.1em;
+      left: 0.1em;
+      border-radius: 2em;
+      transition: left 0.2s ease;
+      border: none;
+      width: 1.2em;
+      height: 1.2em;
+      box-shadow: inset 0.5px -1px 1px rgba(0, 0, 0, 0.35);
+      background: #fff;
+    }
+    &:after {
+      background: transparent;
+      height: calc(100% + 8px);
+      border-radius: 30px;
+      top: -5px;
+      width: calc(100% + 8px);
+      left: -4px;
+      z-index: 0;
+      box-shadow: inset 0px 2px 4px -2px rgba(0, 0, 0, 0.2),
+        0px 1px 2px 0px rgba(151, 151, 151, 0.2);
+    }
+  }
+
+  input:checked + label {
+    background: #4caf50;
+    &:before {
+      left: 1.5em;
+    }
+  }
+}
